video 380 - creating a main class to inherit in other classes
section-38 started
