+++ bienes-raices/classes/Seller.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace App;
+
+class Seller {
+    
+}
